+++ inc/structure/post.php
@@ -95,26 +95,25 @@ if ( ! function_exists( 'storefront_paging_nav' ) ) {
 	 * Display navigation to next/previous set of posts when applicable.
 	 */
 	function storefront_paging_nav() {
-		// Don't print empty markup if there's only one page.
-		if ( $GLOBALS['wp_query']->max_num_pages < 2 ) {
-			return;
-		}
-		?>
-		<nav class="navigation paging-navigation" role="navigation">
-			<h1 class="screen-reader-text"><?php _e( 'Posts navigation', 'storefront' ); ?></h1>
-			<div class="nav-links">
+		global $wp_query;
 
-				<?php if ( get_next_posts_link() ) : ?>
-				<div class="nav-previous"><?php next_posts_link( __( '<span class="meta-nav">&larr;</span> Older posts', 'storefront' ) ); ?></div>
-				<?php endif; ?>
+		$big 		= 999999999; 					// need an unlikely integer
+		$translated = __( 'Page', 'storefront' );	// Supply translatable string
 
-				<?php if ( get_previous_posts_link() ) : ?>
-				<div class="nav-next"><?php previous_posts_link( __( 'Newer posts <span class="meta-nav">&rarr;</span>', 'storefront' ) ); ?></div>
-				<?php endif; ?>
+		echo '<nav class="storefront-pagination">';
 
-			</div><!-- .nav-links -->
-		</nav><!-- .navigation -->
-		<?php
+		echo paginate_links( array(
+			'base'					=> str_replace( $big, '%#%', esc_url( get_pagenum_link( $big ) ) ),
+			'format'				=> '?paged=%#%',
+			'current'				=> max( 1, get_query_var( 'paged' ) ),
+			'prev_text'				=> __( '&larr;', 'storefront' ),
+			'next_text'				=> __( '&rarr;', 'storefront' ),
+			'type'					=> 'list',
+			'total'					=> $wp_query->max_num_pages,
+		    'before_page_number'	=> '<span class="screen-reader-text">' . $translated . ' </span>'
+		) );
+
+		echo '</nav>';
 	}
 }